requirements.txt
streamlit
pyrebase4
requests
google-generativeai
python-dotenv